wacky GUI changes, shapes are not in meshes, better animations, etc

+++ src/lib/GridPoint.java
@@ -1,9 +1,9 @@
 package lib;
 
 public class GridPoint {
-    public double row, col;
+    public int row, col;
 
-    public GridPoint(double row, double col) {
+    public GridPoint(int row, int col) {
         this.row = row;
         this.col = col;
     }
